BAD: duplicated context and passed ID to Form
Here, I gave ExpenseForm the ability to remove Expenses; essentially Edit by delete+duplicate.
But this has the odd behavior of losing track of the id I deleted; making it so I can't edit the same expense twice back to back; I have to edit a different expense first which refreshes the setID.

This introduces small UX errors. So DONT do it this way lol